Removed the function that created a custom http handler, replaced it with creating a custom server class
